+++ src/aspire/covariance/covar2d.py
@@ -607,7 +607,7 @@ class BatchedRotCov2D(RotCov2D):
             ctf_basis_k_sq = ctf_basis_k_t @ ctf_basis_k
             A_mean_k = weight * ctf_basis_k_sq
             A_mean += A_mean_k
-            A_covar_k = np.sqrt(weight) * ctf_basis_k_sq
+            A_covar_k = np.sqrt(weight).astype(self.dtype) * ctf_basis_k_sq
             A_covar[k] = A_covar_k
 
             M_covar += A_covar_k
